refactor: 서비스 의존성 역전 #27
