+++ src/main/java/sogang/capstone/editking/domain/form/FormService.java
@@ -1,75 +1,17 @@
 package sogang.capstone.editking.domain.form;
 
-import java.util.List;
-import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import sogang.capstone.editking.common.exception.ForbiddenException;
-import sogang.capstone.editking.common.response.ErrorCode;
 import sogang.capstone.editking.domain.user.User;
 
-@Service
-@RequiredArgsConstructor
-public class FormService {
+public interface FormService {
 
-    private final FormReader formReader;
-    private final FormStore formStore;
-    private final FormInfoMapper formInfoMapper;
+    public FormInfo.Main registerForm(User user, FormCommand.RegisterForm registerForm);
 
-    @Transactional
-    public FormInfo.Main registerForm(User user, FormCommand.RegisterForm registerForm) {
-        Form form = formStore.store(registerForm.toEntity(user));
-        List<Question> questionList = form.getQuestionList();
-        return formInfoMapper.of(form, questionList);
-    }
+    public void deleteForm(User user, Long formId);
 
-    @Transactional(readOnly = true)
-    private Form validateWriterOfForm(User user, Long formId) {
-        Form form = formReader.getForm(formId);
-        if (form.getUser().getId() != user.getId()) {
-            throw new ForbiddenException(ErrorCode.NOT_WRITER_OF_FORM.getMessage());
-        }
-        return form;
-    }
+    public FormInfo.Main retrieveForm(User user, Long formId);
 
-    @Transactional
-    public void deleteForm(User user, Long formId) {
-        Form form = validateWriterOfForm(user, formId);
-        formStore.delete(form);
-    }
+    public FormInfo.Main editForm(Long formId, User user, FormCommand.EditForm editForm);
 
-    @Transactional(readOnly = true)
-    public FormInfo.Main retrieveForm(User user, Long formId) {
-        Form form = validateWriterOfForm(user, formId);
-        List<Question> questionList = form.getQuestionList();
-        return formInfoMapper.of(form, questionList);
-    }
-
-    @Transactional
-    public FormInfo.Main editForm(Long formId, User user, FormCommand.EditForm editForm) {
-        Form form = validateWriterOfForm(user, formId);
-        form.updatePropertyWith(editForm);
-
-        List<Question> questionList = editForm.getQuestionList().stream().map(Question::new)
-                .collect(Collectors.toList());
-        form.updateQuestionList(questionList);
-
-        return formInfoMapper.of(form, questionList);
-    }
-
-    @Transactional
     public void updateQuestionAndFormStatus(User user, Long formId, Long questionId,
-            FormCommand.UpdateQuestionRequest request) {
-        Form form = validateWriterOfForm(user, formId);
-
-        List<Question> questionList = form.getQuestionList();
-        questionList.stream().filter(question -> question.getIdx() == questionId).findAny()
-                .ifPresent(question -> {
-                            question.updateContent(request.getContent());
-                        }
-                );
-
-        form.updateFormStatus(request.getFormStatus());
-    }
+            FormCommand.UpdateQuestionRequest request);
 }
